first pass on moving device
