+++ src/nanoarrow/device/device_test.cc
