+++ src/nanoarrow/device/metal.cc
