+++ src/nanoarrow/device/metal_test.cc
